requirements.txt
python-telegram-bot[job-queue]==20.7
aiohttp==3.9.3
tenacity==8.4.2
python-dotenv==1.0.1